refactor: rename modules accounts, spending_account to accout, spending_entry
