+++ services/expense-manager-service/app/entities/errors/spending_entry.py
