+++ services/expense-manager-service/app/entities/models/spending_entry.py
