+++ services/expense-manager-service/app/infrastructures/inmemory_db/accounts.py
@@ -2,11 +2,11 @@
 
 from typing import ClassVar
 
-from app.entities.errors.accounts import AccountNotFoundError
+from app.entities.errors.account import AccountNotFoundError
 from app.entities.errors.period import PeriodNotFoundError
-from app.entities.models.accounts import Account
+from app.entities.models.account import Account
 from app.entities.models.period import Period
-from app.entities.repositories.accounts import AccountRepositoryInterface
+from app.entities.repositories.account import AccountRepositoryInterface
 
 
 class AccountRepository(AccountRepositoryInterface):
