+++ services/expense-manager-service/app/infrastructures/postgres_db/models/spending_entry.py
