+++ services/expense-manager-service/app/routers/v1/__init__.py
@@ -2,9 +2,9 @@
 
 from fastapi import APIRouter
 
-from app.routers.v1.endpoints.accounts import account_router
+from app.routers.v1.endpoints.account import account_router
 from app.routers.v1.endpoints.period import period_router
-from app.routers.v1.endpoints.spending_account import router as spending_account_router
+from app.routers.v1.endpoints.spending_entry import router as spending_account_router
 
 router = APIRouter(prefix="/v1")
