+++ services/expense-manager-service/app/routers/v1/schemas/spending_entry.py
