+++ services/expense-manager-service/app/use_cases/errors/spending_entry.py
