+++ services/expense-manager-service/app/use_cases/models/spending_entry.py
